Added comments to public methods, re-ordered ExitComponent recursion.
The reason for calling child.ExitComponent before cleaning up self
references is so that those children still have a reference to Screen.
By disconnecting the parent Component first, the child's self.Screen
gives None.  It might be important (for placement or special drawing
methods) to the child Component to know information about the Screen.