Added c++ tabled version of rotate, but it's clear at this point
that the overhead from ctypes is the limiting factor.  Instead,
I corrected some math errors in the int rotate function in pure python
and I'll just keep using it.  This is the last commit that will include
all of the rotate functions- fast_rot.dll will remain, but only one
rotate function will remain after, which is the python lookup rotation.