Removed extra rotate functions from Util.Math
Now using ONLY int_rotate from trig_tables.  Accurate to 1E-5,
this function is sufficient for the time being.  If/when higher-
precision rotation is needed, I'll probably be looking into other
libraries for that anyway.  So much for the fun with c++ =/
I'll leave the dll in there for now, it was an interesting experiment.