Changed how load/save works for settings.
The Setting class now has methods Load/SaveUsingConfigParser.
The loading and saving is taken care of here, and the Settings class
simply iterates through fields- either in the file, for loading, or
in self.dict for saving- and calls the appropriate method on the
individual setting.

As usual, my try/catch on input is abysmal.  It puts to shame those
that would type catch and simply trace stack.  Silent fail on
load/save error is impressively poor form.