Minor changes to TypedDoubleBuffer, better unittesting for same.
I was using tdb.flip() in testing clear, which isn't ideal.
If I have to use any function of the class in question, it should be
a very, very simple function that is known will never interfere with
a unit test.  Therefore _write_front_buffer and _write_back_buffer
simply call the appropriate appends, and the tests now use these
instead of pushing, flipping, and pushing again.