Renamed some tests, added getter for a copy of EventHandler's listeners
Most of the tests for eventargs ran together, and made sense going
from one to the next, so I kept them together in a single method.

We should have a nice way to inspect an EventHandler's listeners,
but a read-only copy.  Also makes testing easier.  Yay testing!

way
of
inspecting