Purged isinstance from codebase.  Mostly.
There are two instances left:
1) in Wrappers/inject_args, there is a check for basestring, because we
are determining if it is an iterable or a single string, and basing
logic off of that.
2) Structs/TypeCheckedList, because it's... typed.  While I am very
aware of the arguments against isinstance, that's pretty much the only
reason this class exists.  What I WILL do in the next commit is remove
the inheritance from TypeCheckedList.
	For example, TypedLockableList becomes LockableList.