Added PendingAdditions/Removals, improved pending tracking.
Now using a ref count method for additions/removals, as the previous
method would always let removals win over adds. Now if the count for a
pending change is > 0, it gets added.  If a pending count is <= 0,
it gets removed.

PendingAdditions/Removals are a read-only property that allow inspection
of pending changes.