Added test for Settings save.  May consider changing setting init.
Tests have now been added for all of the original material I wanted to
two weeks ago, which is a bit behind schedule but acceptable.
Tests now need to be added for Components and the Logger class, but
I'm not too worried about that.

Regarding default and current index vs value:
Because one is much more likely to pass a value instead of an index for
setting initialization, it seems silly to keep using default_index
or current_index in setting.__init__.  Instead, I think I'll use
default_value, current_value.  This poses a problem if they don't
provide a file or options immediately, so I may change how I store
current/default (invert the model, use value as key, not index).