Added collision code.  Moderate speed, acceptable accuracy.
Not that it's terribly exciting, but I am glad I thought to build-in
an epsilon to each method.  This allows for easy tweaking and debugging.

I'm not sure if it'll go in graphics or collision (probably graphics)
but I'd like to make a collision box/line/circle that can be easily
overlayed, so that someone in debug mode can watch bound sizes change
as they adjust epsilon.

Still need to finish square/circle collisions.