More work in entities; started a full re-write of component.
Component takes after some work in sumplui, although I'm not clever
enough to split out the different pieces (containers and widgets are
merged in my version so far)

I'll have to continue to work on it- preventing an isinstance check has
been a bit demanding.  Also, I've got to be careful to get the up-down
call order for recursives correct, and that always seems to have tripped
me up before...

Rectangle has a change that checks in setattr for an actual change-
if the value gets set to itself, it doesn't flag for a recalc.
This might cost more on the whole with the getattr on every setattr,
but it remains to be seen.