Added some dirty flags and bboxes for collision perf. and readability.
Line gets a bbox so that static collision checks are faster-
	for GUI, line-point, line-rect, point-rect are going to be common,
	and rarely moving, so this will offer some decent gains.
Also, this makes the collision code easier to read and review.

Odd side-effect, points got dirty flags.  This is because when changing
the x or y value of one of the endpoints of a line, you would likely use
line.p1.x += 2, not line.p1 = Point(line.p1.x+2, line.p1.y).
However, line's setattr can only pick up on the second, not the first.
So to correct for that, points are 'dirty' when their x or y change, and
the line checks both points for dirty flags before returning the old
bbox.