Added indexing for StringBuilder, changed dict for LockableList.
As per:
http://wiki.python.org/moin/PythonSpeed/PerformanceTips
I swapped the LockableList to use a defaultdict(int) which means that
when I assign to a key that doesn't exist, it initializes one to 0. 
This allows me to not do an if has_key check each time I add/remove
during a lock.

Added __getitem__ indexing for StringBuilder.

Corrected import problem in Collision.