Removed label, added op overloads for StringBuilder.
Came to a compromise with StringBuilder- it doesn't have most string
methods, but it does have op overloading for common string 
operations, such as: 
	a = "Hello"; a += ", World."
Need to add __bool__ and such, and then when I get home I'll make sure I
didn't miss any major object methods.  Probably not going to add
__hash__, as it is mutable, and giving the hash of the current string
seems dishonest.

Removed label because it was based on the old component, which is gone,
and there's going to be a lot of re-write anyway. Removed unused import
in Events now that Util.contains is gone.