Logger subclasses BufferedWriter.  Collection collision localizes fn.
Collision on collections moves Collision.check into the local namespace
so that for large collections of shapes, the function call overhead is
reduced.  At this point, doing the detection in c would probably be
better, but I don't think this is a premature optimization- most
collections will have quite a few collision objects.

Logger subclasses BufferedWriter, so that for low-priority logging, we
don't have an open/close on every call.  For necessarily immediate
logging, simply use buffer_size = 0.

Still need to do unittest for Logger/BufferedWriter =(