Added bool fn for StringBuilder, BufferedWriter now uses a StringBuilder
Noticed that BufferedWriter's write() method simply did a += on the
string.  Since most of these strings will be long, or build in small
pieces, it's good to take advantage of ''.join() over naive +=.

Since there were some "if self._data:" lines, added __nonzero__ method
to StringBuilder so that it can be used for conditionals.