Refined primes, digits.  Odd git thing going on w/Engine
Git thinks settings and events are updated, but I'm not seeing any diff
in Eclipse.  Ah well.

primes- Tried to write a fast method to extend the list of found primes,
it ended up being leagues slower than a full recalc.  Simply zeroing out
all multiples of found primes is much faster than recalcing a single
prime and carrying it forward, because it has to be redone with each
prime and each set of factors.  In fact, for large ranges on relatively
small ranges (append 50- 100,000 to a list of 2-50) the primes in the
generating list might be needed to cancel own factors, and they'd be
missed using this method.

TL;DR the internet method is much faster, and I need to port it all to c
anyhow.  So full recalc is done when you 'expand' the list.